refactor(eatinorders): infra.jpa 패키지 분리
